hikari-intro-main/requirements.txt
hikari[speedups]~=2.0.0.dev101; python_version < "3.10"
hikari~=2.0.0.dev101; python_version == "3.10"
hikari-lightbulb==2.2.2
git+https://github.com/FasterSpeeding/Tanjun@c1cb6c4aa1b103cda92f6779753867102ceb12d9
uvloop>=0.16; sys_platform != "win32"



aiohttp~=3.8.1
